% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\docType{data}
\name{population}
\alias{population}
\title{Population in USA 1790-2010}
\format{
This data frame contains the observations of 23.
}
\source{
US Bureau of the Census. Web: \url{http://www.census.gov}.
}
\usage{
population
}
\description{
United States population data 1790-2010
}
\examples{
head(population)
}
\keyword{datasets}
