% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\docType{data}
\name{smoking}
\alias{smoking}
\title{Smoking data}
\format{This data frame contains the observations of 91.}
\source{
Max Roser (2016) – 'Smoking'. Published online at OurWorldInData.org. Retrieved from: \url{https://ourworldindata.org/smoking} [Online Resource].

The data for the single countries from the 'International Smoking Statistics' \url{http://www.pnlee.co.uk/ISS.htm}.
}
\usage{
smoking
}
\description{
Sales of cigarettes per adult per day in early industrialized countries, 1920-2010.
}
\examples{
head(smoking)
}
\keyword{datasets}
