% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SSgaussAre}
\alias{SSgaussAre}
\title{Self-Starting nls area of gaussian model}
\usage{
SSgaussAre(input,y0,A,xc,w)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{y0}{offset}

\item{A}{area}

\item{xc}{center}

\item{w}{width}
}
\value{
a numeric vector of the same length as input. It is the value of the expression y0+(A/(w*sqrt(pi/2)))*exp(-2*(x-xc)^2/(w^2)).
}
\description{
Area version of gaussian function.
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
data(smoking)
model <- nlsLM(Germany ~ SSgaussAre(year,y0,A,xc,w),data=smoking)
}
\references{
Origin: \url{http://www.originlab.com/doc/Origin-Help/Gauss-FitFunc}.
}
\seealso{
\code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
