% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SShyper}
\alias{SShyper}
\title{Self-Starting nls hyperbolic model}
\usage{
SShyper(input,a,b)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{a}{estimated model parameter}

\item{b}{estimated model parameter}
}
\value{
a numeric vector of the same length as input. It is the value of the expression (a*x^2)/(x+b).
}
\description{
Computes the hyperbolic function
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
x <- 1:20
y <- (255*x^2)/(x+587)
model <- nls(y ~ SShyper(x,a,b),algorithm = "port")
}
\references{
Goryl A., Jedrzejczyk Z., Kukula K. (Eds.), Osiewalski J., Walkosz A. (2004) "Wprowadzenie do ekonometrii w przykladach i zadaniach", Wydawnictwo Naukowe PWN, Warszawa.

Borkowski B., Dudek H., Szczesny W. (2004) "Ekonometria. Wybrane zagadnienia" , Wydawnictwo Naukowe PWN, Warszawa.
}
\seealso{
\code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
