% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SSlogis2}
\alias{SSlogis2}
\title{Self-Starting nls logistic model}
\usage{
SSlogis2(input,alpha,beta,gamma)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{alpha}{estimated model parameter}

\item{beta}{estimated model parameter}

\item{gamma}{estimated model parameter}
}
\value{
a numeric vector of the same length as input. It is the value of the expression alpha/(1+beta*exp(-gamma*x)).
}
\description{
Computes the logistic function
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
set.seed(46)
x <- 1:15
y <- 1/(1+exp(7-1.5*x))+rnorm(15,0,0.1)
model <- nls(y~SSlogis2(x,alpha,beta,gamma))
}
\references{
Goryl A., Jedrzejczyk Z., Kukula K. (Eds.), Osiewalski J., Walkosz A. (2004) "Wprowadzenie do ekonometrii w przykladach i zadaniach", Wydawnictwo Naukowe PWN, Warszawa.

Borkowski B., Dudek H., Szczesny W. (2004) "Ekonometria. Wybrane zagadnienia" , Wydawnictwo Naukowe PWN, Warszawa.
}
\seealso{
\code{\link{SSlogis}}, \code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
