% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SSlorentz}
\alias{SSlorentz}
\title{Self-Starting nls lorentz model}
\usage{
SSlorentz(input,y0,A,xc,w)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{y0}{offset}

\item{A}{area}

\item{xc}{center}

\item{w}{width}
}
\value{
a numeric vector of the same length as input. It is the value of the expression y0+A*(2*w)/(pi*(4*(x-xc)^2+w^2)).
}
\description{
lorentz peak function.
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
data(smoking)
smoking$t <- 1:nrow(smoking)
model <- nlsLM(Germany ~ SSlorentz(t,y0,A,xc,w),data=smoking)
}
\references{
Origin: \url{http://www.originlab.com/doc/Origin-Help/Lorentz-FitFunc}.
}
\seealso{
\code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
