% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SSpower}
\alias{SSpower}
\title{Self-Starting nls power model}
\usage{
SSpower(input,a,b)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{a}{estimated model parameter}

\item{b}{estimated model parameter}
}
\value{
a numeric vector of the same length as input. It is the value of the expression a*x^b.
}
\description{
Computes the logistic function
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
set.seed(36)
x <- 1:100
y <- 1.25*x^0.45+rnorm(100,0,0.5)
model <- nls(y~SSpower(x,a,b))
}
\references{
Goryl A., Jedrzejczyk Z., Kukula K. (Eds.), Osiewalski J., Walkosz A. (2004) "Wprowadzenie do ekonometrii w przykladach i zadaniach", Wydawnictwo Naukowe PWN, Warszawa.

Borkowski B., Dudek H., Szczesny W. (2004) "Ekonometria. Wybrane zagadnienia" , Wydawnictwo Naukowe PWN, Warszawa.
}
\seealso{
\code{\link[vegan]{SSarrhenius}}, \code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
