% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSpsVoigt2.R
\name{SSpsVoigt2}
\alias{SSpsVoigt2}
\title{Self-Starting nls Pseudo-Voigt model}
\usage{
SSpsVoigt2(input,y0, A, nu, xc, wl, wg)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{y0}{offset}

\item{A}{area}

\item{nu}{profile shape factor}

\item{xc}{center}

\item{wl}{width lorentz}

\item{wg}{width gauss}
}
\value{
a numeric vector of the same length as input. It is the value of the expression y0+A*((nu*2*wl)/(pi*4*(x-xc)^2+wl^2)+(1-nu)*(sqrt(4*log(2))*exp((-4*log(2)*(x-xc)^2)/wg^2))/sqrt(pi*wg)).
}
\description{
lorentz peak function.
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}}.
}
\examples{
data(smoking)
model <- nlsLM(USA~SSpsVoigt2(year,y0, A, nu, xc, wl, wg),data=smoking)
}
\references{
qtiplot: \url{https://www.qtiplot.com/doc/manual-en/x9885.html}.
}
\seealso{
\code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
