% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitCurve.R
\name{SSworking}
\alias{SSworking}
\title{Self-Starting nls Working model}
\usage{
SSworking(input,a,b)
}
\arguments{
\item{input}{a numeric vector of values at which to evaluate the model.}

\item{a}{estimated model parameter}

\item{b}{estimated model parameter}
}
\value{
a numeric vector of the same length as input. It is the value of the expression exp(a+b*(1/x)).
}
\description{
Computes the Working function
}
\note{
This is primarily intended for use in formulae given to the \code{\link{nls}} function or similar functions for example \code{\link{mle2}} and \code{\link{nlsLM}}.
}
\examples{
x <- 1:15
y <- exp(144+3*(1/x))
model <- nlsLM(y~SSworking(x,a,b))
}
\references{
Goryl A., Jedrzejczyk Z., Kukula K. (Eds.), Osiewalski J., Walkosz A. (2004) "Wprowadzenie do ekonometrii w przykladach i zadaniach", Wydawnictwo Naukowe PWN, Warszawa.

Borkowski B., Dudek H., Szczesny W. (2004) "Ekonometria. Wybrane zagadnienia" , Wydawnictwo Naukowe PWN, Warszawa.
}
\seealso{
\code{\link{selfStart}}, \code{\link{getInitial}}
}
\author{
Krzysztof Trajkowski
}
\keyword{function}
